Swap logger to logrus and enable JSON output
Initially I decided not to use a third party logging library as the
needs were quite simple and there was more involved in setup and
maintenance than what it was to just use the stdlib offerings. However,
in #10 it was raised that there is a need to output to JSON. I didn't
particularly think it would be suitable to maintain my own logger with
JSON so I opted to move over to `sirupsen/logrus`.

This project supports text and JSON out of the box and have a bunch of
third party contributions for the formatter[1] and has an easy API
should we need to implement another format.

To use the JSON format, you need to pass `-output-format "json"` to the
executable and all the output from there on will be JSON formatted.

As part of this change, I was able to deprecate some custom code and
still maintain full backwards compatibility. Yay! The only new things
from the log entry is that there is now an additional "level" and
"message". For violation reports, `level` is "info" and `message` is
empty. This isn't huge deal but something to be aware of.

[1]: https://github.com/sirupsen/logrus#formatters